feat: Finalize understanding relationships section
Change the order_by parameter of the get_plot_inputs function to sort by label, absolute or percentage, despite the type of the return data. For example: the function can return churned and stayed lists with percentage values, but it is sorted by the absolute value. This way two charts near each other can be aligned in the labels.

Also organize the variable names and plot the remaining charts for the section.